chore(DataManagementHeader): Moved component
