+++ src/views/data-workflows/gbq-to-gcs/configurations/ListingView.vue
@@ -17,7 +17,7 @@
 </template>
 
 <script>
-import DataManagementHeader from '@/components/data-workflows/DataManagementHeader';
+import DataManagementHeader from '@/components/data-workflows/common/DataManagementHeader';
 import ListingComponent from '@/components/data-workflows/common/listing/ListingComponent';
 
 import TabsItemsMixin from '../tabs-items';
