+++ src/views/data-workflows/storage-to-tables/configurations/ItemView.vue
@@ -6,7 +6,7 @@
 </template>
 
 <script>
-import DataManagementHeader from '../../../../components/data-workflows/DataManagementHeader';
+import DataManagementHeader from '../../../../components/data-workflows/common/DataManagementHeader';
 import RunsItem from '@/components/data-workflows/common/item/ItemComponent';
 
 import TabsItemsMixin from '../tabs-items';
