+++ src/views/data-workflows/storage-to-tables/configurations/ListingView.vue
@@ -21,7 +21,7 @@
 </template>
 
 <script>
-import DataManagementHeader from '../../../../components/data-workflows/DataManagementHeader';
+import DataManagementHeader from '../../../../components/data-workflows/common/DataManagementHeader';
 import ListingComponent from '@/components/data-workflows/common/listing/ListingComponent';
 
 import TabsItemsMixin from '../tabs-items';
