+++ src/views/data-workflows/storage-to-tables/runs/ListingView.vue
@@ -31,8 +31,8 @@
 </template>
 
 <script>
-import DataManagementHeader from '../../../../components/app/headers/DataManagementHeader';
-import ListingComponent from '@/components/data-workflows/common/ListingComponent';
+import DataManagementHeader from '@/components/data-workflows/common/DataManagementHeader';
+import ListingComponent from '@/components/data-workflows/common/listing/ListingComponent';
 import { getStatusColor } from '@/util/data-workflows/run';
 
 import TabsItemsMixin from '../tabs-items';
