+++ src/views/data-workflows/table-to-storage/runs/ItemView.vue
@@ -6,7 +6,7 @@
 </template>
 
 <script>
-import DataManagementHeader from '@/components/data-workflows/DataManagementHeader';
+import DataManagementHeader from '@/components/data-workflows/common/DataManagementHeader';
 import ItemComponent from '@/components/data-workflows/common/item/ItemComponent';
 
 import TabsItemsMixin from '../tabs-items';
