+++ src/views/data-workflows/tables-to-tables/configurations/ListingView.vue
@@ -26,8 +26,8 @@
 </template>
 
 <script>
-import DataManagementHeader from '../../../../components/app/headers/DataManagementHeader';
-import ListingComponent from '@/components/data-workflows/common/ListingComponent';
+import DataManagementHeader from '../../../../components/data-workflows/common/DataManagementHeader';
+import ListingComponent from '@/components/data-workflows/common/listing/ListingComponent';
 import ConfigurationStatus from '@/components/data-workflows/configuration/ConfigurationStatus.vue';
 
 import TabsItemsMixin from '../tabs-items';
